fix(fsm): doc updates and minor refactors
